+++ src/Oro/Bundle/CatalogBundle/Tests/Unit/DependencyInjection/OroCatalogExtensionTest.php
@@ -2,6 +2,7 @@
 
 namespace Oro\Bundle\CatalogBundle\Tests\Unit\DependencyInjection;
 
+use Oro\Bundle\CatalogBundle\DependencyInjection\Configuration;
 use Oro\Bundle\CatalogBundle\DependencyInjection\OroCatalogExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
@@ -25,6 +26,8 @@ class OroCatalogExtensionTest extends \PHPUnit\Framework\TestCase
                         'all_products_page_enabled' => ['value' => false, 'scope' => 'app'],
                         'category_image_placeholder' => ['value' => null, 'scope' => 'app'],
                         'search_autocomplete_max_categories' => ['value' => 2, 'scope' => 'app'],
+                        Configuration::EXCLUDE_CURRENT_BREADCRUMB_ON_ALL_PAGES => ['value' => false, 'scope' => 'app'],
+                        Configuration::REMOVE_SINGLE_BREADCRUMB => ['value' => false, 'scope' => 'app'],
                     ]
                 ]
             ],